backend/requirements.txt
fastapi
pyinstaller
setuptools
uvicorn
python-multipart
torchvision==0.5.0
scipy==1.1.0
numpy==1.18.1
torch==1.4.0
Pillow==7.0.0
opencv-python
